replace validateGraph with assertGraphEquals
